ExRunTime v0.1.0 ExRunTime

Tiny elixir library useful for examining the runtime environment.

Functions

is_kubernetes?(env \\ System.get_env())

is_kubernetes?(%{optional(String.t()) => String.t()}) :: boolean()

Examines the system environment variables to determine whether or not the application is running in Kubernetes.

Returns true or false.

Uses System.get_env() by default but you can pass in your own map.

Examples

iex> ExRunTime.is_kubernetes?(%{"KUBERNETES_POD" => "some-pod"})
true
iex> ExRunTime.is_kubernetes?(%{})
false
iex> ExRunTime.is_kubernetes?()
false

is_production_deploy?(env \\ System.get_env())

is_production_deploy?(%{optional(String.t()) => String.t()}) :: boolean()

Examines the system environment variables to determine whether or not the application is running in the production environment.

Returns true or false.

Uses System.get_env() by default but you can pass in your own map.

Examples

iex> ExRunTime.is_production_deploy?()
false
iex> ExRunTime.is_production_deploy?(%{})
false
iex> ExRunTime.is_production_deploy?(%{"DEPLOYED_ENV" => "test"})
false
iex> ExRunTime.is_production_deploy?(%{"DEPLOYED_ENV" => "prod"})
true

is_test?(env \\ System.get_env())

is_test?(%{optional(String.t()) => String.t()}) :: boolean()

Examines the system environment variables to determine whether or not the application is running in test.

Returns true or false.

Uses System.get_env() by default but you can pass in your own map.

Examples

iex> ExRunTime.is_test?()
false
iex> ExRunTime.is_test?(%{})
false
iex> ExRunTime.is_test?(%{"MIX_ENV" => "test"})
true
iex> ExRunTime.is_test?(%{"MIX_ENV" => "prod"})
false